chore(all): 파일명 수정
admin -> manager
user -> customer
store -> owner
