--- src/main/java/app/domain/admin/service/impl/AdminUserServiceImpl.java
@@ -1,27 +1,0 @@
-package app.domain.admin.service.impl;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
-import app.domain.admin.model.dto.response.GetUserListResponse;
-import app.domain.admin.service.AdminUserService;
-import app.domain.user.model.UserRepository;
-import app.domain.user.model.entity.enums.UserRole;
-import app.global.apiPayload.PagedResponse;
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class AdminUserServiceImpl implements AdminUserService {
-
-	private final UserRepository userRepository;
-
-	@Override
-	public PagedResponse<GetUserListResponse> getAllUsers(Pageable pageable) {
-		Page<GetUserListResponse> page = userRepository.findAllByUserRole(UserRole.CUSTOMER, pageable)
-			.map(GetUserListResponse::from);
-
-		return PagedResponse.from(page);
-	}
-}
