+++ src/main/java/app/domain/customer/model/UserRepository.java
@@ -1,0 +1,19 @@
+package app.domain.customer.model;
+
+import java.util.Optional;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import app.domain.customer.model.entity.User;
+import app.domain.customer.model.entity.enums.UserRole;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+	Optional<User> findByUsername(String userName);
+
+	Page<User> findAllByUserRole(UserRole role, Pageable pageable);
+}
