+++ src/main/java/app/domain/customer/model/entity/UserAddress.java
@@ -1,10 +1,7 @@
-package app.domain.review.model.entity;
+package app.domain.customer.model.entity;
 
 import java.util.UUID;
 
-import app.domain.customer.model.entity.User;
-import app.domain.order.model.entity.Orders;
-import app.domain.owner.model.entity.Store;
 import app.global.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -12,9 +9,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -23,33 +18,30 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "p_review")
+@Table(name = "p_user_address")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class Review extends BaseEntity {
+public class UserAddress extends BaseEntity {
 
 	@Id
 	@GeneratedValue
-	private UUID reviewId;
-
-	@OneToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "b_order_id", nullable = false)
-	private Orders Orders;
+	private UUID addressId;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id", nullable = false)
 	private User user;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "store_id", nullable = false)
-	private Store store;
+	@Column(length = 50)
+	private String alias;
 
 	@Column(nullable = false)
-	private int rating;
+	private String address;
 
-	@Lob
 	@Column(nullable = false)
-	private String context;
+	private String addressDetail;
+
+	@Column(nullable = false)
+	private boolean isDefault = false;
 }
