+++ src/main/java/app/domain/customer/model/entity/enums/UserRole.java
@@ -1,0 +1,13 @@
+package app.domain.customer.model.entity.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum UserRole {
+	CUSTOMER,
+	OWNER,
+	MANAGER,
+	MASTER;
+}
