+++ src/main/java/app/domain/manager/service/AdminUserService.java
@@ -1,0 +1,12 @@
+package app.domain.manager.service;
+
+import org.springframework.data.domain.Pageable;
+
+import app.domain.manager.model.dto.response.GetUserListResponse;
+import app.global.apiPayload.PagedResponse;
+
+public interface AdminUserService {
+
+	PagedResponse<GetUserListResponse> getAllUsers(Pageable pageable);
+
+}
