+++ src/main/java/app/domain/menu/model/entity/Menu.java
@@ -2,7 +2,7 @@ package app.domain.menu.model.entity;
 
 import java.util.UUID;
 
-import app.domain.store.model.entity.Store;
+import app.domain.owner.model.entity.Store;
 import app.global.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
