+++ src/main/java/app/domain/order/model/entity/Orders.java
@@ -2,12 +2,12 @@ package app.domain.order.model.entity;
 
 import java.util.UUID;
 
+import app.domain.customer.model.entity.User;
 import app.domain.order.model.entity.enums.OrderChannel;
 import app.domain.order.model.entity.enums.OrderStatus;
 import app.domain.order.model.entity.enums.PaymentMethod;
 import app.domain.order.model.entity.enums.ReceiptMethod;
-import app.domain.store.model.entity.Store;
-import app.domain.user.model.entity.User;
+import app.domain.owner.model.entity.Store;
 import app.global.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
