+++ src/main/java/app/domain/owner/model/dto/request/StoreApproveRequest.java
@@ -1,0 +1,15 @@
+package app.domain.owner.model.dto.request;
+
+import java.util.UUID;
+
+public record StoreApproveRequest(
+	UUID userId,
+	UUID regionId,
+	String address,
+	String storeName,
+	String desc,
+	String phoneNumber,
+	Long minOrderAmount
+) {
+
+}
