+++ src/main/java/app/domain/owner/model/dto/request/StoreInfoUpdateRequest.java
@@ -1,0 +1,10 @@
+package app.domain.owner.model.dto.request;
+
+public record StoreInfoUpdateRequest(
+	String name,
+	String address,
+	String phoneNumber,
+	String minOrderAmount,
+	String desc
+) {
+}
