+++ src/main/java/app/domain/owner/model/dto/response/StoreApproveResponse.java
@@ -1,0 +1,10 @@
+package app.domain.owner.model.dto.response;
+
+import java.util.UUID;
+
+public record StoreApproveResponse(
+	UUID storeId,
+	String storeApprovalStatus
+) {
+
+}
