+++ src/main/java/app/domain/owner/model/dto/response/StoreInfoUpdateResponse.java
@@ -1,0 +1,8 @@
+package app.domain.owner.model.dto.response;
+
+import java.util.UUID;
+
+public record StoreInfoUpdateResponse(
+	UUID storeId
+) {
+}
