--- src/main/java/app/domain/store/model/dto/request/StoreInfoUpdateRequest.java
@@ -1,10 +1,0 @@
-package app.domain.store.model.dto.request;
-
-public record StoreInfoUpdateRequest(
-	String name,
-	String address,
-	String phoneNumber,
-	String minOrderAmount,
-	String desc
-) {
-}
