--- src/main/java/app/domain/store/model/dto/response/StoreApproveResponse.java
@@ -1,10 +1,0 @@
-package app.domain.store.model.dto.response;
-
-import java.util.UUID;
-
-public record StoreApproveResponse(
-	UUID storeId,
-	String storeApprovalStatus
-) {
-
-}
