--- src/main/java/app/domain/store/model/dto/response/StoreInfoUpdateResponse.java
@@ -1,8 +1,0 @@
-package app.domain.store.model.dto.response;
-
-import java.util.UUID;
-
-public record StoreInfoUpdateResponse(
-	UUID storeId
-) {
-}
