--- src/main/java/app/domain/user/model/entity/UserAddress.java
@@ -1,47 +1,0 @@
-package app.domain.user.model.entity;
-
-import java.util.UUID;
-
-import app.global.entity.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "p_user_address")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class UserAddress extends BaseEntity {
-
-	@Id
-	@GeneratedValue
-	private UUID addressId;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
-
-	@Column(length = 50)
-	private String alias;
-
-	@Column(nullable = false)
-	private String address;
-
-	@Column(nullable = false)
-	private String addressDetail;
-
-	@Column(nullable = false)
-	private boolean isDefault = false;
-}
