--- src/main/java/app/domain/user/model/entity/enums/UserRole.java
@@ -1,13 +1,0 @@
-package app.domain.user.model.entity.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum UserRole {
-	CUSTOMER,
-	OWNER,
-	MANAGER,
-	MASTER;
-}
